+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v 1.5.0
 + Balance Top command (/balancetop, /baltop)
 + Added alias for the set balance command (/setbal)
 + Added a money cap that can be enabled/disabled as well as set from the main config
++ Ability to set default job notifications state. Thanks to @MarkL4YG.
 
 v 1.4.0